feat(ios): add http(s) proxy/scheme
This adds a new URLSchemeHandler which can proxy http(s) requests to external servers.
This is useful for some CORS issues and webview bugs that affect the use of cookies in CORS requests via XHR and fetch.
For that reason cookies will be synced between proxied requests on the native layer and the webview.