fix(ionassethandler.m): fix startPath is getting null (#463)
If Ionic is used inside a framework file, self.basePath is null and stringToLoad contains the full
needed path. Current string concatenation does not respect that case, leading to startPath is null
and so index.html is not loaded. Fix this by setting startPath to empty string instead of null if
base path is null.